requirements/common.txt
pytest==8.0.0
pytest-django==4.8.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.5.0 